move and simplify mina-signer zkapp unit test
